Allow nova_notifier use non-global config
Prototype fix for the issue raised by Julien:

  http://lists.openstack.org/pipermail/openstack-dev/2013-February/thread.html#5861

The nova_notifer code runs in the context of nova, so if we use cfg.CONF
then we are reading options from nova.conf. Instead, we want to read
config options from ceilometer.conf.

To achieve that, all of ceilometers code which is used by nova_notifier
needs to accept a ConfigOpts object created by nova_notifier. The code
needs to both register and read config options from this object.

When we get down to meter_publish, the calls to the oslo RPC code also
need to be passed this ConfigOpts object. To facilitate this, we add
a new rpc.get_impl() function which needs to be proposed to
oslo-incubator. See bug #1131233.

Change-Id: I9bcf7b6fd6d6378a7adb4e0931229f5ce10072cf